Add files via upload
This is a simple console app whereby I'm using the HmlAgilityPack Nuget Package to web scape the common online shop in Kenya jumia.co.ke. In the program.cs file there are two methods namely SeeLaptopsFromPage which takes a page number as a string and returns a list of Laptops on that page. The second method SeeLaptopsFromName takes a name string value, queries the website and returns a list of laptops that match the name.